refactor: replaced messages queue with ExecutorService submits
